pyproject.toml
[project]
name = "nibterm"
version = "0.4.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bino Maiheu", email = "binomaiheu@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jsonpath-ng>=1.6.0",
    "numpy>=1.24.4",
    "paho-mqtt>=2.0.0",
    "pyqtgraph>=0.13.3",
    "pyside6>=6.6.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
nibterm = "nibterm.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.hooks.custom]
path = "build_hooks/version_hook.py"

[tool.hatch.build.targets.wheel.force-include]
"static" = "nibterm/static"

[tool.hatch.build.targets.sdist.force-include]
"static" = "nibterm/static"

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pytest>=8.0",
]
